Fixed gql import in andrjs and added support for gracefull exit in cli
