--- packages/cli/.pnpm-debug.log
@@ -1,16 +1,0 @@
-{
-  "0 debug pnpm:scope": {
-    "selected": 1,
-    "workspacePrefix": "/Users/connorbarr/dev/andromeda/andromeda-js"
-  },
-  "1 error pnpm": {
-    "code": "ERR_PNPM_GIT_NOT_UNCLEAN",
-    "hint": "If you want to disable Git checks on publish, set the \"git-checks\" setting to \"false\", or run again with \"--no-git-checks\".",
-    "err": {
-      "name": "Error",
-      "message": "Unclean working tree. Commit or stash changes first.",
-      "code": "ERR_PNPM_GIT_NOT_UNCLEAN",
-      "stack": "Error: Unclean working tree. Commit or stash changes first.\n    at Object.handler [as publish] (/Users/connorbarr/.node/corepack/pnpm/6.11.0/dist/pnpm.cjs:121004:17)\n    at processTicksAndRejections (node:internal/process/task_queues:96:5)\n    at async /Users/connorbarr/.node/corepack/pnpm/6.11.0/dist/pnpm.cjs:125614:20\n    at async run (/Users/connorbarr/.node/corepack/pnpm/6.11.0/dist/pnpm.cjs:125589:34)\n    at async runPnpm (/Users/connorbarr/.node/corepack/pnpm/6.11.0/dist/pnpm.cjs:125799:5)\n    at async /Users/connorbarr/.node/corepack/pnpm/6.11.0/dist/pnpm.cjs:125791:7"
-    }
-  }
-}
